Move UnboundScript C++ declarations to new header
